+++ experimental/packages/otlp-exporter-base/src/index.ts
@@ -13,7 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './OTLPExporterBase';
 export * from './platform';
-export * from './types';
-export * from './util';
+export { OTLPExporterBase } from './OTLPExporterBase';
+export {
+  OTLPExporterError,
+  OTLPExporterConfigBase,
+  ExportServiceError,
+} from './types';
+export {
+  parseHeaders,
+  appendResourcePathToUrl,
+  appendRootPathToUrlIfNeeded,
+  configureExporterTimeout,
+  invalidTimeout,
+} from './util';